portmixer: Resolve Windows x64 issues.

+++ lib-src/portmixer/src/px_win_common.h
@@ -67,8 +67,8 @@ typedef struct PxInfo
    DWORD       waveID;
 } PxInfo;
 
-int open_mixers(px_mixer *Px, UINT deviceIn, UINT deviceOut);
-int open_ep_mixers(px_mixer *Px, UINT deviceIn, UINT deviceOut);
+int open_mixers(px_mixer *Px, ULONG deviceIn, ULONG deviceOut);
+int open_ep_mixers(px_mixer *Px, ULONG deviceIn, ULONG deviceOut);
 
 #ifdef __cplusplus
 }
